requirements.txt
altair==4.2.0
aniso8601==9.0.1
appdirs==1.4.4
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
contourpy==1.3.2
cryptography==38.0.1
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
ez_setup==0.9
Flask==2.2.2
Flask-API==3.0.post1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
fonttools==4.57.0
gitdb==4.0.10
GitPython==3.1.29
greenlet==1.1.3
idna==3.3
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.17.3
jsonschema-specifications==2024.10.1
jwt==1.3.1
kiwisolver==1.4.8
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.10.1
multitasking==0.0.11
mysql-connector-python==8.0.31
narwhals==1.35.0
numpy==1.23.5
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pipreqs==0.4.11
protobuf==3.20.1
pyarrow==10.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
PyJWT==2.4.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
referencing==0.36.2
requests==2.28.1
rich==12.6.0
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.41
streamlit==1.44.1
streamlit-option-menu==0.4.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
yarg==0.1.9
yfinance==0.1.87
zipp==3.11.0
